refactor: LoadBalancer 분리
